asymmetry {idiogramFISH}	R Documentation

FUNCTIONS asymmetry and asymmetryA2.

Description

asymmetry: calculates karyotype asymmetry A and A2.

asymmetryA2: calculates karyotype asymmetry A2

This functions read a data.frame with columns: shortArmSize and longArmSize

If several species present, use column OTU.

It returns a list with the A and A2 indices

A = Sum (longArm - shortArm / CL) / n

A: Watanabe et al. (1999) asymmetry of karyotype ranging from 0 (symmetric) to 1 (asymmetric)

A2 = sCL / xCL

(s = std dev, CL = chr. length, x = mean) (Romero-Zarco 1986)

related to:

CVCL = A2 * 100

(CV = coeff. var.) (Paszko 2006)

Usage

asymmetry(dfChrSize, asDf = FALSE)

asymmetryA2(dfChrSize)

Arguments

dfChrSize
name of data.frame

asDf
boolean, return d.f. instead of list

Value

list

References

Watanabe K, Yahara T, Denda T, Kosuge K. 1999. Chromosomal evolution in the genus Brachyscome (Asteraceae, Astereae): Statistical tests regarding correlation between changes in karyotype and habit using phylogenetic information. Journal of Plant Research 112: 145-161. 10.1007/PL00013869

A2: Romero-Zarco. 1986. A New Method for Estimating Karyotype Asymmetry. Taxon Vol. 35, No. 3 pp. 526-530

Paszko B. 2006. A critical review and a new proposal of karyotype asymmetry indices. Plant Syst Evol 258:39-48.

See Also

chrbasicdatamono

chrbasicdatamono

chrbasicdataHolo

Examples

asymmetry(dfOfChrSize)
myAlist <- asymmetry(bigdfOfChrSize)
as.data.frame(myAlist)
asymmetryA2(dfOfChrSize)
as.data.frame(asymmetryA2(bigdfOfChrSize))
asymmetryA2(dfChrSizeHolo)
as.data.frame(asymmetryA2(bigdfChrSizeHolo))

[Package idiogramFISH version 2.0.13 Index]