makeVectorNames {idiogramFISH}	R Documentation

makeVectorNames This is an internal function that makes a vector of names

Description

makeVectorNames This is an internal function that makes a vector of names

Usage

makeVectorNames(namedList, attr, selector)

Arguments

namedList
original list

attr
attribute

selector
value of attribute

Value

character vector

[Package idiogramFISH version 2.0.13 Index]