mapXY {idiogramFISH}	R Documentation

mapXY This is an internal function that creates coords for chr. or arms or marks

Description

mapXY This is an internal function that creates coords for chr. or arms or marks

Usage

mapXY(
  start,
  end,
  y,
  yMod,
  x,
  yfactor,
  r2,
  pts_1,
  pts_2,
  pts_3,
  pts_4,
  chrt = FALSE
)

Arguments

start
1st index

end
last index

y
y coords

yMod
y coords modified because of squareness

x
x coords

yfactor
y x factor

r2
radius

pts_1
points of squareness

pts_2
points of squareness

pts_3
points of squareness

pts_4
points of squareness

Value

list

[Package idiogramFISH version 2.0.13 Index]