mapXYCenMimic {idiogramFISH}	R Documentation

mapXYCenMimic This is an internal function that creates coords for chr. or arms or marks

Description

mapXYCenMimic This is an internal function that creates coords for chr. or arms or marks

Usage

mapXYCenMimic(start, end, y, x, yfactor, r2, pts, cenFormat = "triangle")

Arguments

start
1st index

end
last index

y
y coords

yfactor
y x factor

r2
r2ius

pts
points of squareness

cenFormat
boolean

Value

list

[Package idiogramFISH version 2.0.13 Index]