markDistCenGISHfix {idiogramFISH}	R Documentation

markDistCenGISHfix This is an internal function that fixes markDistCen when GISH

Description

markDistCenGISHfix This is an internal function that fixes markDistCen when GISH

Usage

markDistCenGISHfix(
  dfArmGISHInternalMonocen,
  dfChrSizeInternal,
  columnArmSize,
  markDistType,
  listOfdfChromSize,
  addR2 = TRUE
)

Arguments

dfArmGISHInternalMonocen
df of GISH OF ARM

dfChrSizeInternal
d.f. of chr. size

columnArmSize
column rel. to arm.

markDistType
character

Value

data.frame

[Package idiogramFISH version 2.0.13 Index]