markposDFs {idiogramFISH}	R Documentation

Mark Positional data - monocentrics

Description

When several OTUs, some can be monocen. and others holocen. Marks distance for monocen. are measured from cen. and for holocen. from top or bottom depending on param origin. See vignettes.

bigdfOfMarks: Example data for mark position with column OTU

dfOfMarks: Example data for marks' position

dfOfMarks2: Marks' position including cen. marks

humMarkPos: human karyotype bands' (marks) positions, measured from Adler (1994)

allMarksSample: Example data for marks' position

dfAlloParentMarks: Example data for mark position of GISH of monocen.

traspaMarks: T. spathacea (Rhoeo) marks' positions, from Golczyk et al. (2005)

Usage

bigdfOfMarks

dfOfMarks

dfOfMarks2

humMarkPos

allMarksSample

dfAlloParentMarks

traspaMarks

Format

bigdfOfMarks a data.frame with columns:

OTU

OTU, species, mandatory if in dfChrSize

chrName

name of chromosome

markName

name of mark

chrRegion

use p for short arm, q for long arm, and cen for centromeric

markDistCen

distance of mark to centromere (not for cen)

markSize

size of mark (not for cen)

An object of class data.frame with 4 rows and 5 columns.

An object of class data.frame with 6 rows and 5 columns.

An object of class data.frame with 395 rows and 5 columns.

An object of class data.frame with 43 rows and 6 columns.

An object of class data.frame with 16 rows and 4 columns.

An object of class data.frame with 18 rows and 7 columns.

Source

Washington U

References

Adler 1994. Idiogram Album. URL: Washington U.

Golczyk H, Hasterok R, Joachimiak AJ (2005) FISH-aimed karyotyping and characterization of Renner complexes in permanent heterozygote Rhoeo spathacea. Genome 48:145-153.

See Also

markdataholo

plotIdiograms

chrbasicdatamono

dfMarkColor

[Package idiogramFISH version 2.0.13 Index]