namesToColumns {idiogramFISH}	R Documentation

FUNCTION that modifies marks' names into columns

Description

Reads a data.frame with marks' of styles downArrow,upArrow,cM,cMLeft positions. It separates names in columns, avoiding overlap when multiple close names

Exceptionally this function requires the column style in the data.frame of marks' positions.

Returns a data.frame

Usage

 namesToColumns( marksDf, dfChrSize, markType = c("downArrow", "upArrow",
 "cMLeft", "cM"), amountofSpaces = 13, colNumber = 2, protruding = 0.2,
 protrudingInt = 0.5, circularPlot = TRUE, rotation = 0.5,
 defaultStyleMark = "square", orderChr = "size", halfModDown = NA,
 halfModUp = NA, rotatMod = 0 
)

Arguments

marksDf
data.frame with columns: markName,style,markPos

dfChrSize
data.frame, size of chr. Same of plot.

markType
character, use c("downArrow","upArrow","cM","cMLeft") or a subset

amountofSpaces
numeric, number of spaces for each column

colNumber
numeric, number of columns

protruding
numeric, same as plot, minimal protruding for arrow marks, equivalent to cM protruding

protrudingInt
numeric, spacing of columns in terms of width of chr. percent 1 = 100%. Defaults to 0.5

circularPlot
boolean, use TRUE for circular plots. Use FALSE otherwise

rotation
numeric, same as plot, anti-clockwise rotation, defaults to 0.5 which rotates chr. from top to -90 degrees. (-0.5*π )

defaultStyleMark
character, if some data in column style missing fill with this one. Defaults to "square"

orderChr
character, replaces orderBySize - deprecated when "size", sorts chromosomes by total length from the largest to the smallest. "original": preserves d.f. order. "name": sorts alphabetically; "group": sorts by group name

halfModDown
numeric, for circ. plots, when plotting several chromosomes in a circular plot, using a small value 0.05 corrects for alignment problems of downArrows, cMLeft labels. Defaults to NA

halfModUp
numeric, for circ. plots, when plotting several chromosomes in a circular plot, using a small value 0.05 corrects for alignment problems of upArrows, cM labels. Defaults to NA

rotatMod
numeric, for circ. plots, when rotation != 0 (diff.), corrects alignment of labels. Defaults to 0

Value

data.frame

[Package idiogramFISH version 2.0.13 Index]