plotlabelsright {idiogramFISH}	R Documentation

This is an internal function to plot labels to the right, when "aside" for its position

Description

It returns a graphic element with the legends

Usage

plotlabelsright(
  maxx,
  y,
  markLabelSpacer,
  chrWidth,
  dfMarkColorInt,
  allMarkMaxSize,
  normalizeToOne,
  markLabelSize,
  xfactor,
  legendWidth,
  legendHeight,
  n,
  pattern,
  legendYcoord,
  useOneDot,
  dotsAsOval,
  circularPlot
)

Arguments

y
The y axis coordinates of chromosomes

markLabelSpacer
distance from right chr to legend

chrWidth
chr widht

dfMarkColorInt
data.frame of mark characteristics

allMarkMaxSize
maximum size of marks

normalizeToOne
transformation value of karyotype height

markLabelSize
font size of legends

xfactor
relative proportion of x vs y axes

legendWidth
factor to increase width of legend squares and dots

legendHeight
factor to increase height of legend squares and dots

n
numeric, vertices for round parts

pattern,
character, regex to remove from markNames

legendYcoord
numeric modify Y position of legend

useOneDot
boolean, when TRUE plots only one dot

Value

Returns a graphics element

[Package idiogramFISH version 2.0.13 Index]