posCalc {idiogramFISH}	R Documentation

FUNCTION posCalc and fillMarkInfo

Description

calculates position of marks in fraction of (%) chromosome units (0-1)

Usage

posCalc(
  dfMarkPos,
  listOfdfChromSize,
  bToRemove = "",
  markDistType = "beg",
  origin = "b",
  showBandList = FALSE,
  result = "tibble"
)

fillMarkInfo2(dfMarkPos, dfChrSize)

fillMarkInfo(dfMarkPos, dfChrSize, markDistType = "beg", origin = "b")

Arguments

dfMarkPos
data.frame of marks' position

listOfdfChromSize
list (for posCalc) or data.frames of chr. sizes.

bToRemove,
character, bands to remove from calc. of pos.

markDistType
markDistType character, if "cen" = the distance you provided in data.frame (dfMarkPos) column markDistCen or markPos is to the center of the mark, if "beg" = the distance you provided is to the beginning of the mark (Default)

origin,
character, For non-monocentric chr. (for holocentrics only) Use "b" (default) if distance to mark in ("markPos" column in "dfMarkPos") data.frame measured from bottom of chromosome, use "t" for distance to mark from top of chr.

showBandList,
boolean, show row of all bands in tibble, see "result"

result
character, use "tibble" to get results in tibble, "data.frame", or other string results in a list

dfChrSize
data.frame of chr. sizes

Value

list, tibble

data.frame of marks

data.frame of marks

Examples

load(system.file("shinyApps", "iBoard/www/rda/monoholoCS.rda", package = "idiogramFISH"))
load(system.file("shinyApps", "iBoard/www/rda/monoholoMarks.rda", package = "idiogramFISH"))
monoholoMarks2 <- fillMarkInfo(monoholoMarks, monoholoCS)
posCalc(monoholoMarks2, monoholoCS, result = "data.frame")

[Package idiogramFISH version 2.0.13 Index]