robert {idiogramFISH}	R Documentation

FUNCTION to produce a Robertsonian translocation

Description

This function reads a data.frame with chr. sizes chrbasicdatamono and another with marks' positions, markposDFs and gets as arguments two chr. names and two arms, respectively.

It returns a list with two data.frames. One with the chr. size of the resulting translocation and another with the marks' positions for the derivative chr.

Usage

robert(dfChrSize, dfMarkPos, chr1, chr2, arm1, arm2)

Arguments

dfChrSize
name of data.frame of chr. sizes

dfMarkPos
name of data.frame of chr marks' positions

chr1
name of chr.

chr2
name of chr.

arm1
arm of chr1 to be included

arm2
arm of chr2 to be included

Value

list

References

Robertson, W. R. B. (1916). Chromosome studies. I. Taxonomic relationships shown in the chromosomes of Tettigidae and Acrididae: V-shaped chromosomes and their significance in Acrididae, Locustidae, and Gryllidae: chromosomes and variation. Journal of Morphology, 27(2), 179-331.

Examples

data(humChr)
data(humMarkPos)
chrt13q14q <- robert(humChr, humMarkPos, 13, 14, "q", "q")

[Package idiogramFISH version 2.0.13 Index]